# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Satellite Web Service",
        "version": "0.1"
    },
    "paths": {
        "/api/satellites/all_satellites": {
            "get": {
                "tags": [
                    "Satellites"
                ],
                "description": "Get all satellites information",
                "operationId": "6cab67a0f0a48d1af1a58795f096662a",
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/launchpad_satellites": {
            "get": {
                "tags": [
                    "Satellites"
                ],
                "description": "Get satellites by launcpad id",
                "operationId": "9eab4e05d328ed70d7790dffee611ff0",
                "parameters": [
                    {
                        "name": "launchpad_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/categories_satellites": {
            "get": {
                "tags": [
                    "Satellites"
                ],
                "description": "Get satellites by category id",
                "operationId": "7db5570622c0d41df2e5358b99e61709",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/status_satellites": {
            "get": {
                "tags": [
                    "Satellites"
                ],
                "description": "Get satellites by status id",
                "operationId": "c77dea0789ffaca4b6ab7208d23cd02d",
                "parameters": [
                    {
                        "name": "status_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/categories/all_categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "description": "Get all categories information",
                "operationId": "dacf011826cf5fb78b361ed01f03d429",
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/all_status": {
            "get": {
                "tags": [
                    "Status"
                ],
                "description": "Get all status information",
                "operationId": "39a1424d3497719e7dafb2bb22048526",
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/all_launchpads": {
            "get": {
                "tags": [
                    "Launchpads"
                ],
                "description": "Get all launchpads information",
                "operationId": "2dc032fb6d7721371d031c8ea99c558d",
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        }
    }
}